fix non-toplevel imports in several contexts (#475)
