+++ urbanstats/scripts/compute_gpw_data.py
@@ -1,5 +1,5 @@
-from urbanstats.data.gpw import compute_gpw_data_for_shapefile_table
 from urbanstats.geometry.shapefiles.shapefiles_list import shapefiles
+from urbanstats.website_data.table import compute_gpw_data_for_shapefile_table
 
 ys = [y for x, y in shapefiles.items() if y.include_in_gpw]
 for y in ys:
